Move `Async::HTTP::RelativeLocation` to `Async::HTTP::Middleware::LocationRedirector`.
